Remove snap-specific selenium firefox config
On the dev machine I am using, firefox was installed via snap,
meaning that I had to specify the path to the snap geckodriver
for the selenium tests to work. This is fiddly and is now giving
me warning signs for incompatible ffox/gd versions.

Instead of trying to upgrade geckodriver, I've just uninstalled
firefox and reinstalled it via apt. This means that there is no
need for specific selenium for this one machine.